+++ __test__/specs/watchers.spec.js
@@ -13,6 +13,7 @@ describe('Flatpickr watchers', () => {
         value: null,
         config: {
           onChange: () => {
+            // shh..
           }
         }
       }
@@ -27,6 +28,7 @@ describe('Flatpickr watchers', () => {
     const stub = jest.fn();
     wrapper.vm.$on('on-change', stub);
     wrapper.setProps({value: '2017-10-04'});
+
     expect(stub).toHaveBeenCalled();
     stub.mockClear();
   });
@@ -39,7 +41,7 @@ describe('Flatpickr watchers', () => {
     spy.mockReset();
   });
 
-  test('update input value in DOM on value change', () => {
+  test('updates input value in DOM on value change', () => {
     wrapper.setProps({value: '2019-10-04'});
     expect(wrapper.vm.$el.value).toEqual('2019-10-04');
   });